requirements.txt
python-dotenv==1.0.1
requests==2.32.3
flask==3.1.2
hiero-sdk-python==0.1.4
flask-cors==6.0.1
gunicorn
psycopg2-binary
email_validator
flask-sqlalchemy